requirements.txt
Django>=2.1.6
Pillow==5.3.0
django-import-export==1.1.0
django-cleanup
requests
bs4